[14/n][bella-ciao] Remove `LinkageAnalysis` trait and do a bit of general cleanup
