+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/linkage/analysis.rs
@@ -6,85 +6,173 @@ use crate::{
     execution_mode::ExecutionMode,
     static_programmable_transactions::{
         linkage::{
-            config::ResolutionConfig,
-            legacy_linkage,
-            resolution::{ConflictResolution, ResolutionTable, add_and_unify, get_package},
+            config::{LinkageConfig, ResolutionConfig},
+            resolution::{ResolutionTable, VersionConstraint, add_and_unify, get_package},
             resolved_linkage::{ExecutableLinkage, ResolvedLinkage},
         },
         loading::ast::Type,
     },
 };
+use move_binary_format::file_format::Visibility;
 use move_core_types::identifier::IdentStr;
+use move_vm_runtime::validation::verification::ast::Package as VerifiedPackage;
 use sui_protocol_config::ProtocolConfig;
 use sui_types::{
-    base_types::ObjectID, error::ExecutionError, execution_config_utils::to_binary_config,
-    transaction as P,
+    base_types::ObjectID,
+    error::{ExecutionError, ExecutionErrorKind},
+    execution_config_utils::to_binary_config,
 };
 
-pub trait LinkageAnalysis {
-    fn compute_call_linkage(
+#[derive(Debug)]
+pub struct LinkageAnalyzer {
+    internal: ResolutionConfig,
+}
+
+impl LinkageAnalyzer {
+    pub fn new<Mode: ExecutionMode>(
+        protocol_config: &ProtocolConfig,
+    ) -> Result<Self, ExecutionError> {
+        let always_include_system_packages = !Mode::packages_are_predefined();
+        let linkage_config = LinkageConfig::legacy_linkage_settings(always_include_system_packages);
+        let binary_config = to_binary_config(protocol_config);
+        Ok(Self {
+            internal: ResolutionConfig {
+                linkage_config,
+                binary_config,
+            },
+        })
+    }
+
+    pub fn compute_call_linkage(
         &self,
         package: &ObjectID,
         module_name: &IdentStr,
         function_name: &IdentStr,
         type_args: &[Type],
         store: &dyn PackageStore,
-    ) -> Result<ExecutableLinkage, ExecutionError>;
+    ) -> Result<ExecutableLinkage, ExecutionError> {
+        Ok(ExecutableLinkage::new(
+            ResolvedLinkage::from_resolution_table(self.compute_call_linkage_(
+                package,
+                module_name,
+                function_name,
+                type_args,
+                store,
+            )?),
+        ))
+    }
 
-    fn compute_publication_linkage(
+    pub fn compute_publication_linkage(
         &self,
         deps: &[ObjectID],
         store: &dyn PackageStore,
-    ) -> Result<ResolvedLinkage, ExecutionError>;
-
-    fn config(&self) -> &ResolutionConfig;
-}
-
-pub fn linkage_analysis_for_protocol_config<Mode: ExecutionMode>(
-    protocol_config: &ProtocolConfig,
-    _tx: &P::ProgrammableTransaction,
-    store: &dyn PackageStore,
-) -> Result<Box<dyn LinkageAnalysis>, ExecutionError> {
-    Ok(Box::new(legacy_linkage::LegacyLinkage::new(
-        !Mode::packages_are_predefined(),
-        to_binary_config(protocol_config),
-        store,
-    )?))
-}
-
-/// Given a list of object IDs, generate a `ResolvedLinkage` for them.
-/// Since this linkage analysis should only be used for types, all packages are resolved
-/// "upwards" (i.e., later versions of the package are preferred).
-pub fn type_linkage(
-    ids: &[ObjectID],
-    store: &dyn PackageStore,
-) -> Result<ExecutableLinkage, ExecutionError> {
-    let mut resolution_table = ResolutionTable::empty();
-    for id in ids {
-        let pkg = get_package(id, store)?;
-        let transitive_deps = pkg
-            .linkage_table()
-            .values()
-            .map(|version_id| ObjectID::from(*version_id))
-            .collect::<Vec<_>>();
-        let package_id = pkg.version_id().into();
-        add_and_unify(
-            &package_id,
-            store,
-            &mut resolution_table,
-            ConflictResolution::at_least,
-        )?;
-        for object_id in transitive_deps.iter() {
-            add_and_unify(
-                object_id,
-                store,
-                &mut resolution_table,
-                ConflictResolution::at_least,
-            )?;
-        }
+    ) -> Result<ResolvedLinkage, ExecutionError> {
+        Ok(ResolvedLinkage::from_resolution_table(
+            self.compute_publication_linkage_(deps, store)?,
+        ))
     }
 
-    Ok(ExecutableLinkage::new(
-        ResolvedLinkage::from_resolution_table(resolution_table),
-    ))
+    pub fn config(&self) -> &ResolutionConfig {
+        &self.internal
+    }
+
+    fn compute_call_linkage_(
+        &self,
+        package: &ObjectID,
+        module_name: &IdentStr,
+        function_name: &IdentStr,
+        type_args: &[Type],
+        store: &dyn PackageStore,
+    ) -> Result<ResolutionTable, ExecutionError> {
+        let mut resolution_table = self
+            .internal
+            .linkage_config
+            .resolution_table_with_native_packages(store)?;
+
+        fn add_package(
+            object_id: &ObjectID,
+            store: &dyn PackageStore,
+            resolution_table: &mut ResolutionTable,
+            self_resolution_fn: fn(&VerifiedPackage) -> Option<VersionConstraint>,
+            dep_resolution_fn: fn(&VerifiedPackage) -> Option<VersionConstraint>,
+        ) -> Result<(), ExecutionError> {
+            let pkg = get_package(object_id, store)?;
+            let transitive_deps = pkg
+                .linkage_table()
+                .values()
+                .map(|version_id| ObjectID::from(*version_id))
+                .collect::<Vec<_>>();
+            for object_id in transitive_deps.iter() {
+                add_and_unify(object_id, store, resolution_table, dep_resolution_fn)?;
+            }
+            add_and_unify(object_id, store, resolution_table, self_resolution_fn)?;
+            Ok(())
+        }
+
+        let pkg = get_package(package, store)?;
+        let fn_not_found_err = || {
+            ExecutionError::new_with_source(
+                ExecutionErrorKind::FunctionNotFound,
+                format!(
+                    "Could not resolve function '{}' in module {}::{}",
+                    function_name, package, module_name
+                ),
+            )
+        };
+        let fdef = pkg
+            .modules()
+            .iter()
+            .find(|m| m.0.name() == module_name)
+            .ok_or_else(fn_not_found_err)?
+            .1
+            .compiled_module()
+            .find_function_def_by_name(function_name.as_str())
+            .ok_or_else(fn_not_found_err)?;
+
+        let dep_resolution_fn = match fdef.1.visibility {
+            Visibility::Public => VersionConstraint::at_least,
+            Visibility::Private | Visibility::Friend => VersionConstraint::exact,
+        };
+
+        add_package(
+            package,
+            store,
+            &mut resolution_table,
+            VersionConstraint::exact,
+            dep_resolution_fn,
+        )?;
+
+        for type_defining_id in type_args.iter().flat_map(|ty| ty.all_addresses()) {
+            // Type arguments are "at least" constraints
+            add_package(
+                &ObjectID::from(type_defining_id),
+                store,
+                &mut resolution_table,
+                VersionConstraint::at_least,
+                VersionConstraint::at_least,
+            )?;
+        }
+
+        Ok(resolution_table)
+    }
+
+    /// Compute the linkage for a publish or upgrade command. This is a special case because
+    fn compute_publication_linkage_(
+        &self,
+        deps: &[ObjectID],
+        store: &dyn PackageStore,
+    ) -> Result<ResolutionTable, ExecutionError> {
+        let mut resolution_table = self
+            .internal
+            .linkage_config
+            .resolution_table_with_native_packages(store)?;
+        for id in deps {
+            let pkg = get_package(id, store)?;
+            add_and_unify(id, store, &mut resolution_table, VersionConstraint::exact)?;
+            resolution_table
+                .all_versions_resolution_table
+                .insert(pkg.version_id().into(), pkg.original_id().into());
+        }
+        Ok(resolution_table)
+    }
 }
