+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/linkage/mod.rs
@@ -3,6 +3,5 @@
 
 pub mod analysis;
 pub mod config;
-pub mod legacy_linkage;
 pub mod resolution;
 pub mod resolved_linkage;
