+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/linkage/resolution.rs
@@ -15,7 +15,7 @@ use sui_types::{
 
 /// Unifiers. These are used to determine how to unify two packages.
 #[derive(Debug, Clone)]
-pub enum ConflictResolution {
+pub enum VersionConstraint {
     /// An exact constraint unifies as follows:
     /// 1. Exact(a) ~ Exact(b) ==> Exact(a), iff a == b
     /// 2. Exact(a) ~ AtLeast(b) ==> Exact(a), iff a >= b
@@ -28,7 +28,7 @@ pub enum ConflictResolution {
 
 #[derive(Debug, Clone)]
 pub(crate) struct ResolutionTable {
-    pub(crate) resolution_table: BTreeMap<ObjectID, ConflictResolution>,
+    pub(crate) resolution_table: BTreeMap<ObjectID, VersionConstraint>,
     /// For every version of every package that we have seen, a mapping of the ObjectID for that
     /// package to its runtime ID.
     pub(crate) all_versions_resolution_table: BTreeMap<ObjectID, ObjectID>,
@@ -67,30 +67,30 @@ impl ResolutionTable {
     }
 }
 
-impl ConflictResolution {
-    pub fn exact(pkg: &VerifiedPackage) -> Option<ConflictResolution> {
-        Some(ConflictResolution::Exact(
+impl VersionConstraint {
+    pub fn exact(pkg: &VerifiedPackage) -> Option<VersionConstraint> {
+        Some(VersionConstraint::Exact(
             pkg.version(),
             pkg.version_id().into(),
         ))
     }
 
-    pub fn at_least(pkg: &VerifiedPackage) -> Option<ConflictResolution> {
-        Some(ConflictResolution::AtLeast(
+    pub fn at_least(pkg: &VerifiedPackage) -> Option<VersionConstraint> {
+        Some(VersionConstraint::AtLeast(
             pkg.version(),
             pkg.version_id().into(),
         ))
     }
 
     #[allow(dead_code)]
-    pub fn no_constraint(_pkg: &MovePackage) -> Option<ConflictResolution> {
+    pub fn no_constraint(_pkg: &MovePackage) -> Option<VersionConstraint> {
         None
     }
 
-    pub fn unify(&self, other: &ConflictResolution) -> Result<ConflictResolution, ExecutionError> {
+    pub fn unify(&self, other: &VersionConstraint) -> Result<VersionConstraint, ExecutionError> {
         match (&self, other) {
             // If we have two exact resolutions, they must be the same.
-            (ConflictResolution::Exact(sv, self_id), ConflictResolution::Exact(ov, other_id)) => {
+            (VersionConstraint::Exact(sv, self_id), VersionConstraint::Exact(ov, other_id)) => {
                 if self_id != other_id || sv != ov {
                     Err(ExecutionError::new_with_source(
                         ExecutionErrorKind::InvalidLinkage,
@@ -101,13 +101,13 @@ impl ConflictResolution {
                         ),
                     ))
                 } else {
-                    Ok(ConflictResolution::Exact(*sv, *self_id))
+                    Ok(VersionConstraint::Exact(*sv, *self_id))
                 }
             }
             // Take the max if you have two at least resolutions.
             (
-                ConflictResolution::AtLeast(self_version, sid),
-                ConflictResolution::AtLeast(other_version, oid),
+                VersionConstraint::AtLeast(self_version, sid),
+                VersionConstraint::AtLeast(other_version, oid),
             ) => {
                 let id = if self_version > other_version {
                     *sid
@@ -115,7 +115,7 @@ impl ConflictResolution {
                     *oid
                 };
 
-                Ok(ConflictResolution::AtLeast(
+                Ok(VersionConstraint::AtLeast(
                     *self_version.max(other_version),
                     id,
                 ))
@@ -123,12 +123,12 @@ impl ConflictResolution {
             // If you unify an exact and an at least, the exact must be greater than or equal to
             // the at least. It unifies to an exact.
             (
-                ConflictResolution::Exact(exact_version, exact_id),
-                ConflictResolution::AtLeast(at_least_version, at_least_id),
+                VersionConstraint::Exact(exact_version, exact_id),
+                VersionConstraint::AtLeast(at_least_version, at_least_id),
             )
             | (
-                ConflictResolution::AtLeast(at_least_version, at_least_id),
-                ConflictResolution::Exact(exact_version, exact_id),
+                VersionConstraint::AtLeast(at_least_version, at_least_id),
+                VersionConstraint::Exact(exact_version, exact_id),
             ) => {
                 if exact_version < at_least_version {
                     return Err(ExecutionError::new_with_source(
@@ -142,7 +142,7 @@ impl ConflictResolution {
                     ));
                 }
 
-                Ok(ConflictResolution::Exact(*exact_version, *exact_id))
+                Ok(VersionConstraint::Exact(*exact_version, *exact_id))
             }
         }
     }
@@ -168,7 +168,7 @@ pub(crate) fn add_and_unify(
     object_id: &ObjectID,
     store: &dyn PackageStore,
     resolution_table: &mut ResolutionTable,
-    resolution_fn: fn(&VerifiedPackage) -> Option<ConflictResolution>,
+    resolution_fn: fn(&VerifiedPackage) -> Option<VersionConstraint>,
 ) -> Result<(), ExecutionError> {
     let package = get_package(object_id, store)?;
